requirements.txt
asgiref==3.9.1
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
chardet==4.0.0
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==5.2.5
fail2ban==0.11.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iotop==0.6
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-debian===0.1.43ubuntu1
python-dotenv==1.1.1
python-magic==0.4.24
PyYAML==5.4.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.4
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.14.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0


django-filter
djangorestframework